Scope issue fixed, flux debugging ongoing
I have fixed an issue related to the Julian behaviour of scopes. In order to update variables in a for loop which are initialized outside, a "let" clause is needed. The problem with the flux computation is still ongoing, right now I'm debugging the diffusivity() function.  I have also updated the Manifest and Project files after updating to Julia 1.6.