Ongoing: SIA UDEs with errors
First efforts to create a SIA UDE. Two things need to be developed so far: (1) train the UDE only at the end of a simulation period (which should be shortened), in order to simplify the problem; and (2) compute the H matrix inside loss() in order to compute the loss function and backpropagate with Zygote. Follow the #TODO comments.