Staggered grid keeps original size(H)
The code following ParallelStencil style has been updated in order to keep the original matrix size and avoid solving it only for the interior of the matrix. It still needs to be adjusted in order to perform the adjoints, whether by allowing to modify H with Zygote.buffer or by storing intermediate steps.