MWE fixed for optimization of D
An adaptation of the NN's architecture fixed the training for the NN of D. I have updated the MWE and the UDE_D notebook accordingly. I have also tried the fix for iceflow.jl but it did not do the trick. For now, I've modified iceflow.jl to work with a scalar A instead of a matrix, in order to simply things.