[WIP] Weird behaviour of UDE training of A
We need to investigate why the loss function is coupled to the absolute predicted value of A. I have added new plots to track the evolution of the training. This should make it easier to understand what is going on. I think there is an issue with the reference dataset we are generating.