[WIP] MWE with Argentiere and still debugging issue with ice_dynamics
Right now the MWE with Argentiere is too slow to work, we need to wait for the bugs to be fixed. Meanwhile, the ice_dynamics_DiffEqs.jl is failing with a serialization problem. It needs to be debugged to see why it's not finding the same behaviour as the MWE.