UDE training with noise
I have moved the parallelization from an EnsembleProblem to a classic pmap. So far this allows to run PDE simulations with a progress bar. I still haven't made it work with UDEs because of problems with adjoints. I have also added random (seeded) noise to the reference dataset, and the UDE seems to still correctly learn even in the presence of noise