UDE training with ice velocities
Some refactoring has been done throughout the code. Now we can store the reference dataset as a dictionary, which is more practical. Moreover, now we can store and use the ice velocities (V) for the loss function. This will be the preferred solution, to be used with the dataset from Millan et al. (2022).