DiffEqFlux deprecated
I have updated the SciML package environment, for which DiffEqFlux has been deprecated for our needs. I migrated the code to use SciMLSensitivity and Optimization. I have also tested QuadratureAdjoint with Zygote and it works this time. Enzyme is still failing, which wil require more exploration and possibly an MWE.