New MB models and workflow  (#109)
* [WIP] Towards inversions without solver

Trying to fetch ITS_LIVE data, which is only available for very few glaciers, so it might be worth waiting for Romain's data. I'm still implementing the retrieval of Glathida ice thickness data. There is no OGGM task for that yet, so we have to implement it manually.

* [WIP] Rheology inversion with Millan 22

Still need to correctly retrieve Glathida data and project it to the glacier mesh.

* Initial version of rheology inversions working

So far the problem works with ADAM, but there's an issue with BFGS. We need to develop the training further, since now it quickly reaches a local minima

* Rheology inversions with 100 glaciers

Inversions working with a larger pool of glaciers. Now the code can automatically filter missing glaciers and continue only with the ones with data. Some parts might require running the code again after failure. Since there are some try/catch statements and some read/write of files, this should be verified that it correctly works in another machine prior to testing.

* [WIP] Strange error on forward PDE with MB

Strange behaviour during reference simulation. There's a leak on the accummulation area of glaciers, where there is positive MB. There seems to be a mistake in the way the MB is added, but I haven't identified what is breaking the simulations with respect to previous versions.

* MB fixed using DiscreteCallback

Now the MB is correctly added and in a more efficient way by using a DiscreteCallback

* [WIP] UDE with MB using DiscreteCallback NOT working

* Ice rheology inversions for A w/ reference dataset working

* Batching working for ice rheology inversions

First version of batching working. The loss function and the training still needs exploration to become efficient.

* [WIP] Weird behaviour of batching, to be investigated

Workflow incomplete, so far only applied for diffusivity_inversion_experiment.jl. To be propagated to the other workflows once completed. Mostly the batching scheme and the new callbacks that work with batching.

* Diffusivity inversion experiment script added

* [WIP] Inversions not working with batches.

* change loss function for V

* Fix problem with initial conditions. Inversion working with ADAM and no further customization.

* remove diagnosis plots

* Pre-merge commit

* [WIP] Batching and Float64 for UDEs. Still not working

First attempt to use batching and move to Float64 for UDEs for the toy model. Something is making the pullback fail, it needs to be properly debugged.

* [WIP] UDEs failing with NaN values in t

* version working with UDE

* Stable codebase: PDEs and UDEs working

This branch focuses on the UDE part for the paper. The UE inversions on real data should work, but that workflow has not been checked yet.

* Code clean-up and merging code from DiscreteCallbacks

* util functions improved to avoid memroy allocation on PDEs)

I have added new version of the basic matrix operations for staggered grids for solving the PDE. I have only implemented them in some of the operations. The remaining ones require changing the size of the initialized matrices. Since this is not urgent I'm leaving this for later.

* PDE test fixed

PDE solve test updated and fixed to work with new code.

* Long bug fixed + test working

I have included Facu's fix from the error in the context tuple. I have also updated the test comparing the UDE and PDE and now they converge within reasonable limits. We just need to check if the UDE trains correctly now.

* Test working with velocities

Now the test works with H and V. The training of the UDE seems to run but it's very slow. We need to check that it converges.

* Benchmark updated

Benchmark updated to look for an efficient solver for the UDEs.

* Test file updated and improved benchmark

I have committed the correct test file (before it was the wrong one). I have also included sensealg testing in the benchmark, on top of testing multiple solvers in parallel.

* [WIP] Fixing AD+Diff

* [WIP] Finite differences implemented but not converging

I have implemented both in-place and out-of-place version. So far there seems to be a problem with the gradients. I have also added a script to benchmark memory usage. To be used later.

* [WIP] New climate and mass balance workflow

New more memory efficent way to compute the climate series and the MB. Also introducing structs for MB models in order to start implementing this in a clean way.

* [WIP] New MB models implemented in DiscreteCallback

Now the MB is computed in a DiscreteCallback. A new type for MB models has been created, which is the start for a new API to correctly integrate the MB component.

* PDE and UDE working with new MB component

The UDE with finite differences still needs to be tested. Tests need to be redone to accommodate new workflow. This branch needs to be merged with Facu's latest changes with the finite differences workflow.

* Memory issue resolved. Workflow with new MB models working

The memory issue has been temporariliy solved. It was coming from the OGGM multiprocessing. Now simulations can be run with really reasonable amounts of RAM. The test has been updated and the workflow is working with the new climate data and MB models.

* New MB models working with finite differences + code clean-up.

The new workflow of climate and MB models has been correctly added to the workflow of finite differences. Moreover, I have removed all the old code of climate data processing and random MB.

* Running the GC in each worker in UDE to tame the RAM

* Dependency clean-up and start of merge with finite differences

I have cleaned up package dependencies that were no longer used. I have also merged the code with the new changes from finite differences, but so far it doesn't seem to work. To be revised

* small changes in the diff method

---------

Co-authored-by: Facundo Sapienza <fsapienza@berkeley.edu>