First prototype of new API for predictions
First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.