First prototype of new API (#125)
* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API