New architecture using Sleipnir, Huginn and Muninn (#139)
* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API

* First prototype of new API (#125)

* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API

* Plot scripts updated

* New API prototype (#127)

* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API

* First prototype of new API (#125)

* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API

* Plot scripts updated

* MB model much more efficient

I improved the way the climate is handled by the MB model, making it much more efficient without almost any memory allocation.

* New architecture up to date with latest changes of Sleipnir, Huginn and Muninn

All the redundant files have been removed, and now everything is imported from the sublibraries. The workflow works for predictions, but no tests are implemented yet, since they'll have to cover Inversion and FunctionalInversion, which are still to be coded.

* CI updated with Julia certificates and CI for any branch

* Update Huginn version

* Dependencies fixed

* Huginn updated

* CI updated with fixed SSL certificates for Python

* Issue with multiprocessing fixed

* Multiprocessing for OGGM fixed

* Huginn version changed

* Bump version and update Huginn

* Test mode added for tests