[WIP] Debugging Reverse after merge with ForwardDiff setup (#150)
* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API

* First prototype of new API (#125)

* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API

* Plot scripts updated

* New API prototype (#127)

* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API

* First prototype of new API (#125)

* UDEs working with experiment with different MB

Fixed a few bugs in the UDE workflow. The UDEs work suprisingly well with a different MB model. This means that using surface velocities is really a well proxy, which is robust to changes in surface elevation.

* New composite types for parameters created

* Updated mass balance model types

* New types and structs for new API

The basic structure for the new API is ready. Now we need to slowly couple it with the old workflows.

* First prototype of new API for predictions

First prototype more or less ready. The test script is running, but I need to optimize the way Results are saved in a file. Right now it's way too slow and heavy. We need to trim that down and save it as netCDF. The tests are still under construction to fit the new API. And of course, I still need to design the hardest part, the UDE.

* [WIP] Debugging PDE behaviour of new API

* Plot scripts updated

* MB model much more efficient

I improved the way the climate is handled by the MB model, making it much more efficient without almost any memory allocation.

* CompatHelper: bump compat for Polynomials to 4, (keep existing compat) (#131)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* CompatHelper: bump compat for Flux to 0.14, (keep existing compat) (#129)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
Co-authored-by: Jordi Bolibar <jordi.bolibar@gmail.com>

* Update publication information

* New architecture up to date with latest changes of Sleipnir, Huginn and Muninn

All the redundant files have been removed, and now everything is imported from the sublibraries. The workflow works for predictions, but no tests are implemented yet, since they'll have to cover Inversion and FunctionalInversion, which are still to be coded.

* 🔄 synced file(s) with ODINN-SciML/Sleipnir.jl (#130)

* 🔄 synced local 'environment.yml' with remote 'environment.yml'

* 🔄 synced local '.gitignore' with remote '.gitignore'

* CompatHelper: bump compat for AbbreviatedStackTraces to 0.2, (keep existing compat) (#132)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* CompatHelper: add new compat entry for Downloads at version 1, (keep existing compat) (#133)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* CompatHelper: add new compat entry for Statistics at version 1, (keep existing compat) (#134)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* CompatHelper: bump compat for Makie to 0.20, (keep existing compat) (#135)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* CompatHelper: bump compat for CairoMakie to 0.11, (keep existing compat) (#136)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* CI updated with Julia certificates and CI for any branch

* Update Huginn version

* Dependencies fixed

* Huginn updated

* CI updated with fixed SSL certificates for Python

* Issue with multiprocessing fixed

* Multiprocessing for OGGM fixed

* Huginn version changed

* Bump version and update Huginn

* CompatHelper: bump compat for GR to 0.73, (keep existing compat) (#138)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* Test mode added for tests

* 🔄 synced file(s) with ODINN-SciML/Sleipnir.jl (#143)

* 🔄 synced local 'environment.yml' with remote 'environment.yml'

* 🔄 synced local '.github/workflows/CI.yml' with remote '.github/workflows/CI.yml'

* CompatHelper: bump compat for OptimizationOptimisers to 0.2, (keep existing compat) (#142)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* CompatHelper: bump compat for GeoMakie to 0.6, (keep existing compat) (#140)

Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>

* [WIP] Reverse diff for functional inversions

Work in progress commit. Still errors due to problems with mutable structs. We need to decide to go either for ReverseDiff with PreallocationTools or with Enzyme.

* Constrain Huginn version

* Updated Project.toml

* [WIP] Debugging Reverse after merge with ForwardDiff setup

Back to the point of debugging where a choice needs to be made between ReverseDiff and Enzyme. Now this aslo includes the changes from the classic inversions using ForwardDiff.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
Co-authored-by: Facundo Sapienza <fsapienza@berkeley.edu>