Multiprocessing and precompilation improved (#269)
* Multiprocessing and precompilation improved

I have improved a few things in terms of precompilation, making sure we only rely on the necessary dependencies. Moreover, I have re-implemented the forward run in the inversion for the ground truth to fully run in parallel, and I have verified that the training runs in parallel. There is also a (commented) implementation of package compilation, which is way too slow to be useful, but which we might be interested in fixing in the future.

* Fix small issues

* Add Interpolations.jl again

* Add GeoStats.jl again and fix namespace collision with Polynomials.jl

* [WIP] Some gradient tests not passing with new multiprocessing functions. To be investigated, as it seems there is a small numerical degradation, with the values not being really off.